scraper/requirements.txt
mysql-connector-python
flask==3.0.0
flask_cors==4.0.0
flasgger